Fixed Warnings and a Bool issue

+++ thomas/ThomasCore/src/thomas/Physics.cpp
@@ -57,14 +57,14 @@ namespace thomas
 				return;
 			}
 		}
-		int size = s_rigidBodies.size();
+		int size = (int)s_rigidBodies.size();
 		s_rigidBodies.push_back(rigidBody);
 		s_world->addRigidBody(rigidBody, GetCollisionGroupBit(rigidBody->m_gameObject->GetLayer()), GetCollisionMask(rigidBody->m_gameObject->GetLayer()));
 		
 	}
 	bool Physics::RemoveRigidBody(object::component::Rigidbody * rigidBody)
 	{
-		int size = s_rigidBodies.size();
+		int size = (int)s_rigidBodies.size();
 		bool found = false;
 		for (unsigned i = 0; i < s_rigidBodies.size(); ++i)
 		{
